# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catering Business API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  category: string;
  price: number;
  /** Number of people one unit serves */
  servingSize: number;
  /** e.g. "tray", "dozen", "per person" */
  unit: string;
  imageUrl?: string | null;
  allergens: string[];
  available: boolean;
  prepTime?: string | null;
  createdAt: string;
}

export interface CreateMenuItemBody {
  name: string;
  description: string;
  category: string;
  price: number;
  servingSize: number;
  unit: string;
  imageUrl?: string | null;
  allergens: string[];
  available: boolean;
  prepTime?: string | null;
}

export interface UpdateMenuItemBody {
  name?: string;
  description?: string;
  category?: string;
  price?: number;
  servingSize?: number;
  unit?: string;
  imageUrl?: string | null;
  allergens?: string[];
  available?: boolean;
  prepTime?: string | null;
}

export interface BlackoutDate {
  id: number;
  date: string;
  reason?: string | null;
  createdAt: string;
}

export interface CreateBlackoutDateBody {
  date: string;
  reason?: string | null;
}

export interface AvailabilityResponse {
  available: boolean;
  blackoutDates: string[];
  suggestedDates: string[];
}

export interface CartItem {
  id: number;
  menuItemId: number;
  menuItem: MenuItem;
  quantity: number;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
}

export interface AddToCartBody {
  sessionId: string;
  menuItemId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  sessionId: string;
  quantity: number;
}

export interface PlanItem {
  id: number;
  menuItemId: number;
  menuItem: MenuItem;
}

export interface Plan {
  sessionId: string;
  items: PlanItem[];
}

export interface AddToPlanBody {
  sessionId: string;
  menuItemId: number;
}

export interface OrderItem {
  id: number;
  menuItemId: number;
  menuItemName: string;
  price: number;
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  preparing: "preparing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  sessionId: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  eventDate?: string | null;
  eventType?: string | null;
  guestCount?: number | null;
  serviceStyle?: string | null;
  deliveryNotes?: string | null;
  status: OrderStatus;
  total: number;
  items: OrderItem[];
  createdAt: string;
}

export interface CreateOrderBody {
  sessionId: string;
  customerName: string;
  customerEmail: string;
  customerPhone?: string | null;
  eventDate?: string | null;
  eventType?: string | null;
  guestCount?: number | null;
  serviceStyle?: string | null;
  deliveryNotes?: string | null;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  preparing: "preparing",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export type ChatHistoryMessageRole =
  (typeof ChatHistoryMessageRole)[keyof typeof ChatHistoryMessageRole];

export const ChatHistoryMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatHistoryMessage {
  role: ChatHistoryMessageRole;
  content: string;
}

export interface ChatMessageBody {
  sessionId: string;
  message: string;
  history?: ChatHistoryMessage[];
}

export interface SuggestItemsBody {
  guestCount: number;
  serviceStyle: string;
  preferences?: string[];
}

export interface ItemSuggestion {
  menuItemId: number;
  menuItem: MenuItem;
  recommendedQuantity: number;
  reason: string;
}

export interface SuggestItemsResponse {
  suggestions: ItemSuggestion[];
}

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  totalMenuItems: number;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListMenuItemsParams = {
  category?: string;
  available?: boolean;
};

export type CheckAvailabilityParams = {
  startDate: string;
  endDate: string;
};

export type GetCartParams = {
  sessionId: string;
};

export type GetPlanParams = {
  sessionId: string;
};