# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Catering Business API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/menu:
    get:
      operationId: adminListMenuItems
      tags: [admin]
      summary: List all menu items (admin)
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [admin]
      summary: Create a new menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /admin/menu/{id}:
    put:
      operationId: updateMenuItem
      tags: [admin]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMenuItem
      tags: [admin]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Event / Calendar endpoints
  /events/availability:
    get:
      operationId: checkAvailability
      tags: [events]
      summary: Check availability for a date range
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Availability info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
  /admin/blackout-dates:
    get:
      operationId: listBlackoutDates
      tags: [admin]
      summary: List all blackout dates
      responses:
        "200":
          description: Blackout dates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlackoutDate"
    post:
      operationId: createBlackoutDate
      tags: [admin]
      summary: Add a blackout date
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlackoutDateBody"
      responses:
        "201":
          description: Created blackout date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlackoutDate"
  /admin/blackout-dates/{id}:
    delete:
      operationId: deleteBlackoutDate
      tags: [admin]
      summary: Remove a blackout date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Cart & Orders
  /cart:
    get:
      operationId: getCart
      tags: [orders]
      summary: Get current cart (by session)
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [orders]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [orders]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [orders]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the catering AI agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageBody"
      responses:
        "200":
          description: SSE stream of AI response
          content:
            text/event-stream: {}
  /chat/suggest-items:
    post:
      operationId: suggestMenuItems
      tags: [chat]
      summary: Get AI menu item suggestions based on event details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestItemsBody"
      responses:
        "200":
          description: Suggested menu items with quantities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestItemsResponse"

  # Admin stats
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        servingSize:
          type: integer
          description: Number of people one unit serves
        unit:
          type: string
          description: e.g. "tray", "dozen", "per person"
        imageUrl:
          type: string
          nullable: true
        allergens:
          type: array
          items:
            type: string
        available:
          type: boolean
        prepTime:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - category
        - price
        - servingSize
        - unit
        - allergens
        - available
        - createdAt
    CreateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        servingSize:
          type: integer
        unit:
          type: string
        imageUrl:
          type: string
          nullable: true
        allergens:
          type: array
          items:
            type: string
        available:
          type: boolean
        prepTime:
          type: string
          nullable: true
      required:
        - name
        - description
        - category
        - price
        - servingSize
        - unit
        - allergens
        - available
    UpdateMenuItemBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        servingSize:
          type: integer
        unit:
          type: string
        imageUrl:
          type: string
          nullable: true
        allergens:
          type: array
          items:
            type: string
        available:
          type: boolean
        prepTime:
          type: string
          nullable: true
    BlackoutDate:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - createdAt
    CreateBlackoutDateBody:
      type: object
      properties:
        date:
          type: string
          format: date
        reason:
          type: string
          nullable: true
      required:
        - date
    AvailabilityResponse:
      type: object
      properties:
        available:
          type: boolean
        blackoutDates:
          type: array
          items:
            type: string
            format: date
        suggestedDates:
          type: array
          items:
            type: string
            format: date
      required:
        - available
        - blackoutDates
        - suggestedDates
    CartItem:
      type: object
      properties:
        id:
          type: integer
        menuItemId:
          type: integer
        menuItem:
          $ref: "#/components/schemas/MenuItem"
        quantity:
          type: integer
      required:
        - id
        - menuItemId
        - menuItem
        - quantity
    Cart:
      type: object
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
      required:
        - sessionId
        - items
        - total
    AddToCartBody:
      type: object
      properties:
        sessionId:
          type: string
        menuItemId:
          type: integer
        quantity:
          type: integer
      required:
        - sessionId
        - menuItemId
        - quantity
    UpdateCartItemBody:
      type: object
      properties:
        sessionId:
          type: string
        quantity:
          type: integer
      required:
        - sessionId
        - quantity
...
    ChatMessageBody:
      type: object
      properties:
        sessionId:
          type: string
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatHistoryMessage"
      required:
        - sessionId
        - message
    ChatHistoryMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content
    SuggestItemsBody:
      type: object
      properties:
        guestCount:
          type: integer
        serviceStyle:
          type: string
        preferences:
          type: array
          items:
            type: string
      required:
        - guestCount
        - serviceStyle
    SuggestItemsResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/ItemSuggestion"
      required:
        - suggestions
    ItemSuggestion:
      type: object
      properties:
        menuItemId:
          type: integer
        menuItem:
          $ref: "#/components/schemas/MenuItem"
        recommendedQuantity:
          type: integer
        reason:
          type: string
      required:
        - menuItemId
        - menuItem
        - recommendedQuantity
        - reason
...