# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Catering Business API
 * OpenAPI spec version: 0.1.0
 */
import type { MenuItem } from "./menuItem";

export interface CartItem {
  id: number;
  menuItemId: number;
  menuItem: MenuItem;
  quantity: number;
}